make db the last default parameter always
If db isn't the last default parameter of a function and you wanted to use another parameter with a default that came after the db parameter, then you'd have to fill in the defaultdb parameter to use that other parameter. Since this api seems to push the idea of setting a defaultdb then ensuring that you don't have to fill in the db parameter would be the priority. So other parameters with defaults (such as options) should be listed before the db parameter with its default. This is what this commit does.